Refactor: move char utils into scripts::chars
